fix(events): refactor Koji events
