+++ packit_service/worker/events/koji/__init__.py
@@ -1,0 +1,2 @@
+# Copyright Contributors to the Packit project.
+# SPDX-License-Identifier: MIT
