+++ packit_service/worker/events/koji/base.py
@@ -1,5 +1,6 @@
 # Copyright Contributors to the Packit project.
 # SPDX-License-Identifier: MIT
+
 import logging
 from typing import Optional, Union
 
@@ -11,110 +12,20 @@ from packit.utils.koji_helper import KojiHelper
 from packit_service.config import PackageConfigGetter
 from packit_service.constants import KojiBuildState, KojiTaskState
 from packit_service.models import (
-    AbstractProjectObjectDbType,
     GitBranchModel,
-    KojiBuildTargetModel,
-    ProjectEventModel,
     ProjectReleaseModel,
     PullRequestModel,
 )
-from packit_service.worker.events.abstract import Result
 from packit_service.worker.events.event import (
     use_for_job_config_trigger,
 )
+from packit_service.worker.events.koji.abstract import KojiEvent
 
 logger = logging.getLogger(__name__)
 
 
-class AbstractKojiEvent(Result):
-    def __init__(
-        self,
-        task_id: int,
-        rpm_build_task_ids: Optional[dict[str, int]] = None,
-        start_time: Optional[Union[int, float, str]] = None,
-        completion_time: Optional[Union[int, float, str]] = None,
-    ):
-        super().__init__()
-        self.task_id = task_id
-        # dictionary with archs and IDs, e.g. {"x86_64": 123}
-        self.rpm_build_task_ids = rpm_build_task_ids
-        self.start_time: Optional[Union[int, float, str]] = start_time
-        self.completion_time: Optional[Union[int, float, str]] = completion_time
-
-        # Lazy properties
-        self._target: Optional[str] = None
-        self._build_model: Optional[KojiBuildTargetModel] = None
-        self._build_model_searched = False
-
-    @property
-    def build_model(self) -> Optional[KojiBuildTargetModel]:
-        if not self._build_model_searched and not self._build_model:
-            self._build_model = KojiBuildTargetModel.get_by_task_id(
-                task_id=self.task_id,
-            )
-            self._build_model_searched = True
-        return self._build_model
-
-    def get_db_project_object(self) -> Optional[AbstractProjectObjectDbType]:
-        return self.build_model.get_project_event_object() if self.build_model else None
-
-    def get_db_project_event(self) -> Optional[ProjectEventModel]:
-        return self.build_model.get_project_event_model() if self.build_model else None
-
-    @property
-    def target(self) -> Optional[str]:
-        if not self._target and self.build_model:
-            self._target = self.build_model.target
-        return self._target
-
-    @staticmethod
-    def get_koji_rpm_build_web_url(
-        rpm_build_task_id: int,
-        koji_web_url: str = "https://koji.fedoraproject.org",
-    ) -> str:
-        """
-        Constructs the web URL for the given Koji task.
-        You can redefine the Koji instance using the one defined in the service config.
-        """
-        return f"{koji_web_url}/koji/taskinfo?taskID={rpm_build_task_id}"
-
-    @staticmethod
-    def get_koji_build_logs_url(
-        rpm_build_task_id: int,
-        koji_logs_url: str = "https://kojipkgs.fedoraproject.org",
-    ) -> str:
-        """
-        Constructs the log URL for the given Koji task.
-        You can redefine the Koji instance using the one defined in the service config.
-        """
-        return (
-            f"{koji_logs_url}//work/tasks/{rpm_build_task_id % 10000}/{rpm_build_task_id}/build.log"
-        )
-
-    def get_koji_build_rpm_tasks_logs_urls(
-        self,
-        koji_logs_url: str = "https://kojipkgs.fedoraproject.org",
-    ) -> dict[str, str]:
-        """
-        Constructs the log URLs for all RPM subtasks of the Koji task.
-        """
-        return {
-            arch: AbstractKojiEvent.get_koji_build_logs_url(
-                rpm_build_task_id=rpm_build_task_id,
-                koji_logs_url=koji_logs_url,
-            )
-            for arch, rpm_build_task_id in self.rpm_build_task_ids.items()
-        }
-
-    def get_dict(self, default_dict: Optional[dict] = None) -> dict:
-        result = super().get_dict()
-        result.pop("_build_model")
-        result.pop("_build_model_searched")
-        return result
-
-
 @use_for_job_config_trigger(trigger_type=JobConfigTriggerType.commit)
-class KojiBuildEvent(AbstractKojiEvent):
+class Build(KojiEvent):
     def __init__(
         self,
         build_id: int,
@@ -198,8 +109,8 @@ class KojiBuildEvent(AbstractKojiEvent):
         return result
 
     @classmethod
-    def from_event_dict(cls, event: dict):
-        return KojiBuildEvent(
+    def from_event_dict(cls, event: dict) -> "Build":
+        return Build(
             build_id=event.get("build_id"),
             state=KojiBuildState(raw_new) if (raw_new := event.get("state")) else None,
             old_state=(KojiBuildState(raw_old) if (raw_old := event.get("old_state")) else None),
@@ -221,7 +132,7 @@ class KojiBuildEvent(AbstractKojiEvent):
         )
 
 
-class KojiTaskEvent(AbstractKojiEvent):
+class Task(KojiEvent):
     """
     Used for scratch builds.
     """
@@ -293,8 +204,8 @@ class KojiTaskEvent(AbstractKojiEvent):
         return self._identifier
 
     @classmethod
-    def from_event_dict(cls, event: dict):
-        return KojiTaskEvent(
+    def from_event_dict(cls, event: dict) -> "Task":
+        return Task(
             task_id=event.get("task_id"),
             state=KojiTaskState(event.get("state")) if event.get("state") else None,
             old_state=(KojiTaskState(event.get("old_state")) if event.get("old_state") else None),
@@ -327,7 +238,7 @@ class KojiTaskEvent(AbstractKojiEvent):
 
 
 @use_for_job_config_trigger(trigger_type=JobConfigTriggerType.koji_build)
-class KojiBuildTagEvent(AbstractKojiEvent):
+class BuildTag(KojiEvent):
     _koji_helper: Optional[KojiHelper] = None
 
     def __init__(
@@ -373,8 +284,8 @@ class KojiBuildTagEvent(AbstractKojiEvent):
         return f"{self.package_name}-{self.version}-{self.release}"
 
     @classmethod
-    def from_event_dict(cls, event: dict) -> "KojiBuildTagEvent":
-        return KojiBuildTagEvent(
+    def from_event_dict(cls, event: dict) -> "BuildTag":
+        return BuildTag(
             build_id=event.get("build_id"),
             tag_id=event.get("tag_id"),
             tag_name=event.get("tag_name"),
