+++ packit_service/worker/handlers/distgit.py
@@ -88,7 +88,7 @@ from packit_service.worker.events import (
     ReleaseEvent,
     ReleaseGitlabEvent,
 )
-from packit_service.worker.events.koji import KojiBuildTagEvent
+from packit_service.worker.events.koji.base import BuildTag as KojiBuildTagEvent
 from packit_service.worker.handlers.abstract import (
     JobHandler,
     RetriableJobHandler,
