+++ packit_service/worker/handlers/koji.py
@@ -58,7 +58,8 @@ from packit_service.worker.events import (
     ReleaseEvent,
     ReleaseGitlabEvent,
 )
-from packit_service.worker.events.koji import KojiBuildEvent, KojiBuildTagEvent
+from packit_service.worker.events.koji.base import Build as KojiBuildEvent
+from packit_service.worker.events.koji.base import BuildTag as KojiBuildTagEvent
 from packit_service.worker.handlers.abstract import (
     JobHandler,
     TaskName,
